+++ changelog.md
@@ -1,4 +1,5 @@
 - Profiling events are no longer emitted unless a performance profile is being gathered.
+- Changed profiler shortcut to CTRL + SHIFT + F10.
 
 v0.6.8